Add all datagrid filters tests
